+++ lib/pleroma/config/deprecation_warnings.ex
@@ -25,21 +25,19 @@ defmodule Pleroma.Config.DeprecationWarnings do
   def check_skip_thread_containment do
     # The default in config/config.exs is "true" since 593b8b1e6a8502cca9bf5559b8bec86f172bbecb
     # but when the default is retrieved in code the fallback is still "false"
-    uses_thread_visibility_filtering = !Config.get([:instance, :skip_thread_containment], false)
+    uses_thread_visibility_filtering = Config.get([:instance, :skip_thread_containment]) != nil
 
     if uses_thread_visibility_filtering do
       Logger.warning("""
-      !!!DEPRECATION WARNING!!!
-      Your config is explicitly enabling thread-based visibility containment by setting the below:
-      ```
+      !!!CONFIG ERROR!!!
       config :pleroma, :instance, skip_thread_containment: false
-      ```
-
-      This feature comes with a very high performance overhead and is considered for removal.
-      If you actually need or strongly prefer keeping it, speak up NOW(!) by filing a ticket at
-         https://akkoma.dev/AkkomaGang/akkoma/issues
-      Complaints only after the removal happened are much less likely to have any effect.
+      was removed after a deprecation window during which no complaints were raised.
+      You should drop this setting from your config.
       """)
+
+      :error
+    else
+      :ok
     end
   end