+++ lib/pleroma/web/nodeinfo/nodeinfo.ex
@@ -78,7 +78,6 @@ defmodule Pleroma.Web.Nodeinfo.Nodeinfo do
         "mailerEnabled" => Config.get([Pleroma.Emails.Mailer, :enabled], false),
         "features" => features,
         "restrictedNicknames" => Config.get([Pleroma.User, :restricted_nicknames]),
-        "skipThreadContainment" => Config.get([:instance, :skip_thread_containment], false),
         "privilegedStaff" => Config.get([:instance, :privileged_staff]),
         "localBubbleInstances" => Config.get([:instance, :local_bubble], []),
         "publicTimelineVisibility" => %{