+++ test/pleroma/web/activity_pub/activity_pub_test.exs
@@ -1149,58 +1149,6 @@ defmodule Pleroma.Web.ActivityPub.ActivityPubTest do
     end
   end
 
-  describe "timeline post-processing" do
-    test "it filters broken threads" do
-      user1 = insert(:user)
-      user2 = insert(:user)
-      user3 = insert(:user)
-
-      {:ok, user1, user3} = User.follow(user1, user3)
-      assert User.following?(user1, user3)
-
-      {:ok, user2, user3} = User.follow(user2, user3)
-      assert User.following?(user2, user3)
-
-      {:ok, user3, user2} = User.follow(user3, user2)
-      assert User.following?(user3, user2)
-
-      {:ok, public_activity} = CommonAPI.post(user3, %{status: "hi 1"})
-
-      {:ok, private_activity_1} = CommonAPI.post(user3, %{status: "hi 2", visibility: "private"})
-
-      {:ok, private_activity_2} =
-        CommonAPI.post(user2, %{
-          status: "hi 3",
-          visibility: "private",
-          in_reply_to_status_id: private_activity_1.id
-        })
-
-      {:ok, private_activity_3} =
-        CommonAPI.post(user3, %{
-          status: "hi 4",
-          visibility: "private",
-          in_reply_to_status_id: private_activity_2.id
-        })
-
-      activities =
-        ActivityPub.fetch_activities([user1.ap_id | User.following(user1)])
-        |> Enum.map(fn a -> a.id end)
-
-      private_activity_1 = Activity.get_by_ap_id_with_object(private_activity_1.data["id"])
-
-      assert [public_activity.id, private_activity_1.id, private_activity_3.id] == activities
-
-      assert length(activities) == 3
-
-      activities =
-        ActivityPub.fetch_activities([user1.ap_id | User.following(user1)], %{user: user1})
-        |> Enum.map(fn a -> a.id end)
-
-      assert [public_activity.id, private_activity_1.id] == activities
-      assert length(activities) == 2
-    end
-  end
-
   describe "flag/1" do
     setup do
       reporter = insert(:user)